Fix minor warnings with MSVC

+++ IlmBase/Half/half.h
@@ -2,9 +2,9 @@
 //
 // Copyright (c) 2002, Industrial Light & Magic, a division of Lucas
 // Digital Ltd. LLC
-// 
+//
 // All rights reserved.
-// 
+//
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided that the following conditions are
 // met:
@@ -16,8 +16,8 @@
 // distribution.
 // *       Neither the name of Industrial Light & Magic nor the names of
 // its contributors may be used to endorse or promote products derived
-// from this software without specific prior written permission. 
-// 
+// from this software without specific prior written permission.
+//
 // THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
 // "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
 // LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
@@ -278,7 +278,7 @@ HALF_EXPORT void        printBits   (char  c[35], float f);
 #define HALF_MANT_DIG	11		// Number of digits in mantissa
 					// (significand + hidden leading 1)
 
-// 
+//
 // floor( (HALF_MANT_DIG - 1) * log10(2) ) => 3.01... -> 3
 #define HALF_DIG	3		// Number of base 10 digits that
 					// can be represented without change
@@ -319,9 +319,9 @@ HALF_EXPORT void        printBits   (char  c[35], float f);
 //	floating point number, whose bits are arranged as follows:
 //
 //	    31 (msb)
-//	    | 
+//	    |
 //	    | 30     23
-//	    | |      | 
+//	    | |      |
 //	    | |      | 22                    0 (lsb)
 //	    | |      | |                     |
 //	    X XXXXXXXX XXXXXXXXXXXXXXXXXXXXXXX
@@ -365,7 +365,7 @@ HALF_EXPORT void        printBits   (char  c[35], float f);
 //	Here is the bit-layout for a half number, h:
 //
 //	    15 (msb)
-//	    | 
+//	    |
 //	    | 14  10
 //	    | |   |
 //	    | |   | 9        0 (lsb)
@@ -471,7 +471,7 @@ half::half (float f)
 	    //
 
 	    int m = x.i & 0x007fffff;
-	    _h = e + ((m + 0x00000fff + ((m >> 13) & 1)) >> 13);
+	    _h = (unsigned short) (e + ((m + 0x00000fff + ((m >> 13) & 1)) >> 13));
 	}
 	else
 	{
@@ -559,7 +559,7 @@ half::round (unsigned int n) const
 // Other inline functions
 //-----------------------
 
-inline half	
+inline half
 half::operator - () const
 {
     half h;
@@ -640,7 +640,7 @@ half::operator /= (float f)
 }
 
 
-inline bool	
+inline bool
 half::isFinite () const
 {
     unsigned short e = (_h >> 10) & 0x001f;
@@ -690,7 +690,7 @@ half::isInfinity () const
 }
 
 
-inline bool	
+inline bool
 half::isNegative () const
 {
     return (_h & 0x8000) != 0;
